Theme, Gulp and Layout Updates/Authentication Fix
Why?
===
Needed to look better and the authentication request was being triggered with each
request.

What Changed?
===
Made it so that the authentication request for a token wasn't being made with each request,
cleaned up some code and updated the UI to look a bit better. Also updated the bootstrap
theme to change the primary color.